fix room reference
It fixes a problem that it ignore a room number from another script.
It is because hubot use envelope.room to identify channel as standard,
but hubot-kandan use envelope.room.id for it.

A room number originated from other scripts or hubot itself
is ignored and always set default '1'.

A root cause was;
'envelop' is introduced in hubot commit dc27aca6fcbd0
in 12 Jan 2013.
But hubot-kandan put name 'user' for it as same as kandan.

There are room.id and user.id in kandan, and is difference with hubot.
This commit aims to use as same termnology as hubot.

envelope.user - The User
envelope.room - The room identifier

Signed-off-by: Hiroshi Miura <miurahr@linux.com>